Eslint and Stylelint errors fixed
