+++ src/js/beam.js
@@ -5,7 +5,8 @@ shot when the user press the spacebar
 
 /****************************************************** */
 
-// eslint-disable-next-line
+import Phaser from 'phaser';
+
 class Beam extends Phaser.GameObjects.Sprite {
   constructor(scene) {
     const x = scene.player.x + 7;
@@ -14,9 +15,7 @@ class Beam extends Phaser.GameObjects.Sprite {
     super(scene, x, y, 'beam');
 
     scene.add.existing(this);
-    this.play('beam_anim');
     scene.physics.world.enableBody(this);
-    this.body.velocity.y = -250;
     scene.projectiles.add(this);
   }
