+++ src/js/create.js
@@ -51,6 +51,15 @@ const create = (() => {
     return button;
   };
 
+  const inputNameField = () => {
+    const div = document.createElement('div');
+    div.setAttribute('id', 'playerNameBox');
+    const input = document.createElement('input');
+    input.setAttribute('id', 'playerName');
+    div.appendChild(input);
+    document.body.appendChild(div);
+  };
+
   const cursors = (scene) => scene.input.keyboard.createCursorKeys();
   return {
     text,
@@ -58,6 +67,7 @@ const create = (() => {
     musicConfiguration,
     cursors,
     button,
+    inputNameField,
   };
 })();
